Tesselation, searching, homogeneity in Bboxes
-boundingbox.tesselate now works. It will split a region into 2
subregions along its majoral axis if it isn't homogeneous.
-boundingbox.getRegion now finds the lowest level region which contains
the given point.
-Deleted CheckPoints and simplified checkEdges. I'll optimize later when
the slicer is dead slow but working.
-replaced checkInside with isInside so i could tab complete, and made it
actually update self.inside in all cases. I haven't yet written a
procedure to update the .inside flag on all bottom nodes, but it should
be trivial since I already have getBottom.
-getBottom outputs a list of all the regions which lack children, either
because they are homogeneous or because they are max depth.
-.volume returns volume