Unfinished state, due to be redone
I'm rebooting this project and switching to a new approach. I plan to
shift every point in the object mesh by the Z value of its XY
coordinates on the slicing surface.  I can then slice the distorted
object into flat layers which is much less complex. Lastly, I shift
every point in each layer downward by the Z value of its XY coords in
the slicing surface. This results in layers with the contours of the
slicing surface which add up to the original object.