Fixed bug mentioned in previous commit
tri.plane_intersect definition was very outdated, and was effectively
redefining the direction vectors for the 3 edges in the triangle. It was
defining them as arrays of shape [3,], which lack a starting point and
therefore broke the plane.vector_intersect method used in
tri.plane_intersect causing it to output floats instead of points.